+++ packages/fiori-elements-writer/src/data/defaults.ts
@@ -1,7 +1,15 @@
 import { OdataVersion } from '@sap-ux/odata-service-writer';
 import type { OdataService } from '@sap-ux/odata-service-writer';
 import readPkgUp from 'read-pkg-up';
-import type { ALPSettings, ALPSettingsV2, ALPSettingsV4, FioriElementsApp, Template } from '../types';
+import type {
+    ALPSettings,
+    ALPSettingsV2,
+    ALPSettingsV4,
+    FioriElementsApp,
+    LROPSettings,
+    Template,
+    WorklistSettings
+} from '../types';
 import { TableSelectionMode, TableType, TemplateType } from '../types';
 import { getBaseComponent, getUi5Libs, TemplateTypeAttributes } from './templateAttributes';
 
@@ -40,6 +48,17 @@ export function setDefaultTemplateSettings<T extends {}>(template: Template<T>, 
                 smartVariantManagement: alpSettingsv2.smartVariantManagement
             });
             return templateSettings;
+        }
+    } else if (template.type === TemplateType.ListReportObjectPage || template.type === TemplateType.Worklist) {
+        const tableSettings: WorklistSettings | LROPSettings = template.settings as unknown as
+            | WorklistSettings
+            | LROPSettings;
+        Object.assign(templateSettings, {
+            tableType: tableSettings.tableType ?? TableType.RESPONSIVE // Overrides the UI5 default: ''
+        });
+
+        if (tableSettings.tableType !== TableType.TREE) {
+            delete tableSettings.hierarchyQualifier;
         }
     }
     return templateSettings;
@@ -89,7 +108,7 @@ export function setAppDefaults<T>(feApp: FioriElementsApp<T>): FioriElementsApp<
             feApp.ui5.version ?? TemplateTypeAttributes[feApp.template.type].minimumUi5Version[feApp.service.version]!;
     }
 
-    // if not explictly disabled, enable the SAP Fiori tools
+    // if not explicitly disabled, enable the SAP Fiori tools
     feApp.appOptions = feApp.appOptions ?? {};
     if (feApp.appOptions.sapux !== false) {
         feApp.appOptions.sapux = true;