+++ packages/preview-middleware/src/base/flex.ts
@@ -1,45 +1,10 @@
 import type { Logger } from '@sap-ux/logger';
 import type { ReaderCollection } from '@ui5/fs';
-import { existsSync, mkdirSync, readdirSync, unlinkSync, writeFileSync } from 'fs';
+import type { Editor } from 'mem-fs-editor';
+import { existsSync, readdirSync, unlinkSync } from 'fs';
 import { join, parse } from 'path';
+import type { CommonChangeProperties } from '@sap-ux/adp-tooling';
 import { TelemetryReporter } from './telemetry-reporter';
-
-/**
- * Structure of a flex change.
- */
-export interface FlexChange {
-    [key: string]: string | object | undefined;
-    changeType: string;
-    reference: string;
-    moduleName?: string;
-    content?: {
-        codeRef?: string;
-        fragmentPath?: string;
-    };
-}
-
-/**
- * Map of change type specific correction functions.
- */
-const moduleNameContentMap: { [key: string]: (change: FlexChange) => string } = {
-    codeExt: (change) => (change.content?.codeRef ?? '').replace('.js', ''),
-    addXML: (change) => change.content?.fragmentPath ?? ''
-};
-
-/**
- * Sets the moduleName property of the provided change to also support old changes with newer UI5 versions.
- *
- * @param change change to be fixed
- * @param logger logger instance
- */
-function tryFixChange(change: FlexChange, logger: Logger) {
-    try {
-        const prefix = change.reference.replace(/\./g, '/');
-        change.moduleName = `${prefix}/changes/${moduleNameContentMap[change.changeType](change)}`;
-    } catch (error) {
-        logger.warn('Could not fix missing module name.');
-    }
-}
 
 /**
  * Read changes from the file system and return them.
@@ -48,15 +13,15 @@ function tryFixChange(change: FlexChange, logger: Logger) {
  * @param logger logger instance
  * @returns object with the file name as key and the file content as value
  */
-export async function readChanges(project: ReaderCollection, logger: Logger): Promise<Record<string, FlexChange>> {
-    const changes: Record<string, FlexChange> = {};
+export async function readChanges(
+    project: ReaderCollection,
+    logger: Logger
+): Promise<Record<string, CommonChangeProperties>> {
+    const changes: Record<string, CommonChangeProperties> = {};
     const files = await project.byGlob('/**/changes/*.*');
     for (const file of files) {
         try {
-            const change = JSON.parse(await file.getString()) as FlexChange;
-            if (moduleNameContentMap[change.changeType] && !change.moduleName) {
-                tryFixChange(change, logger);
-            }
+            const change = JSON.parse(await file.getString()) as CommonChangeProperties;
             changes[`sap.ui.fl.${parse(file.getName()).name}`] = change;
             logger.debug(`Read change from ${file.getPath()}`);
         } catch (error) {
@@ -73,6 +38,7 @@ export async function readChanges(project: ReaderCollection, logger: Logger): Pr
  * @param data.fileName file name that is required for a valid change
  * @param data.fileType file type that is required
  * @param webappPath path to the webapp folder
+ * @param fs mem-fs editor
  * @param logger logger instance
  * @returns object with success flag and optional message
  */
@@ -86,6 +52,7 @@ export function writeChange(
         selector?: { type?: string };
     },
     webappPath: string,
+    fs: Editor,
     logger: Logger,
     reporter?: TelemetryReporter
 ): { success: boolean; message?: string } {
@@ -93,12 +60,8 @@ export function writeChange(
     const fileType = data.fileType;
     if (fileName && fileType) {
         logger.debug(`Write change ${fileName}.${fileType}`);
-        const path = join(webappPath, 'changes');
-        if (!existsSync(path)) {
-            mkdirSync(path);
-        }
-        const filePath = join(path, fileName + '.' + fileType);
-        writeFileSync(filePath, JSON.stringify(data, null, 2));
+        const filePath = join(webappPath, 'changes', fileName + '.' + fileType);
+        fs.writeJSON(filePath, data);
         const telemetryData = {
             category: 'Save',
             changeType: data.changeType ?? 'Unknown',