+++ examples/simple-generator/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/generator-simple-fe
 
+## 0.3.41
+
+### Patch Changes
+
+-   Updated dependencies [31f5027c]
+    -   @sap-ux/axios-extension@1.11.7
+    -   @sap-ux/system-access@0.3.21
+
+## 0.3.40
+
+### Patch Changes
+
+-   Updated dependencies [8b227bb4]
+    -   @sap-ux/fiori-elements-writer@0.18.15
+
 ## 0.3.39
 
 ### Patch Changes