+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @sap-ux/adp-tooling
 
+## 0.9.20
+
+### Patch Changes
+
+-   00cf3025: Alternative approach to have a consistent save for XML Fragments
+
+## 0.9.19
+
+### Patch Changes
+
+-   Updated dependencies [4389c528]
+    -   @sap-ux/project-access@1.19.11
+    -   @sap-ux/axios-extension@1.11.7
+    -   @sap-ux/system-access@0.3.21
+
+## 0.9.18
+
+### Patch Changes
+
+-   Updated dependencies [31f5027c]
+    -   @sap-ux/axios-extension@1.11.7
+    -   @sap-ux/system-access@0.3.21
+
+## 0.9.17
+
+### Patch Changes
+
+-   fc55fd6b: Component Usages writer produces incorrect content of a change
+
+## 0.9.16
+
+### Patch Changes
+
+-   Updated dependencies [f8e16120]
+    -   @sap-ux/project-access@1.19.10
+    -   @sap-ux/axios-extension@1.11.6
+    -   @sap-ux/system-access@0.3.20
+
 ## 0.9.15
 
 ### Patch Changes