+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/telemetry
 
+## 0.4.18
+
+### Patch Changes
+
+-   Updated dependencies [4389c528]
+    -   @sap-ux/project-access@1.19.11
+
+## 0.4.17
+
+### Patch Changes
+
+-   Updated dependencies [f8e16120]
+    -   @sap-ux/project-access@1.19.10
+
 ## 0.4.16
 
 ### Patch Changes